Adopt BEC in EntityBender
