+++ src/main/java/com/crowsofwar/avatar/common/entity/mob/EntityBender.java
@@ -16,10 +16,9 @@
 */
 package com.crowsofwar.avatar.common.entity.mob;
 
-import com.crowsofwar.avatar.common.data.BendingData;
 import com.crowsofwar.avatar.common.data.Bender;
+import com.crowsofwar.avatar.common.data.BenderEntityComponent;
 import net.minecraft.entity.EntityCreature;
-import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.world.World;
 
@@ -28,64 +27,38 @@ import net.minecraft.world.World;
  * 
  * @author CrowsOfWar
  */
-public abstract class EntityBender extends EntityCreature implements Bender {
+public abstract class EntityBender extends EntityCreature {
 	
-	private BendingData data;
+	private final BenderEntityComponent bender;
 	
 	/**
 	 * @param world
 	 */
 	public EntityBender(World world) {
 		super(world);
-		data = new BendingData(category -> {}, () -> {}); // mc automatically saves entity
+		bender = new BenderEntityComponent(this);
 	}
 	
 	@Override
 	public void readEntityFromNBT(NBTTagCompound nbt) {
 		super.readEntityFromNBT(nbt);
-		data.readFromNbt(nbt);
+		bender.getData().readFromNbt(nbt);
 	}
 	
 	@Override
 	public void writeEntityToNBT(NBTTagCompound nbt) {
 		super.writeEntityToNBT(nbt);
-		data.writeToNbt(nbt);
+		bender.getData().writeToNbt(nbt);
 	}
 	
 	@Override
 	public void onUpdate() {
 		super.onUpdate();
-		data.decrementCooldown();
+		bender.getData().decrementCooldown();
 	}
-	
-	@Override
-	public EntityLivingBase getEntity() {
-		return this;
+
+	public Bender getBender() {
+		return bender;
 	}
-	
-	@Override
-	public BendingData getData() {
-		return data;
-	}
-	
-	@Override
-	public boolean isCreativeMode() {
-		return false;
-	}
-	
-	@Override
-	public boolean isFlying() {
-		return false;
-	}
-	
-	@Override
-	public boolean isPlayer() {
-		return false;
-	}
-	
-	@Override
-	public boolean consumeWaterLevel(int amount) {
-		return false;
-	}
-	
+
 }
